Reduce circular dependencies
